requirements.txt
streamlit
groq
duckduckgo-search
yfinance
pandas
matplotlib
mplfinance
python-dotenv
fpdf2